fix(debug-table): split debug_dispatch.hpp so AVR macros never reach user code
generated_debug.cpp emits `&g_config.INSTANCE0.<varname>` for every leaf
in the project's debug table.  On AVR targets, that file included
`debug_dispatch.hpp` which `#include <avr/pgmspace.h>` to get PROGMEM
and the `pgm_read_*_far` accessors.  pgmspace.h transitively pulls
`<avr/io.h>`, which defines a slew of register names as preprocessor
macros — `SP` → `_SFR_IO16(0x3D)` → `(*(volatile uint16_t *)(0x3D))`,
plus SREG, OCR0A, TCNT0, etc.

IEC 61131-3 function blocks routinely use those identifiers as field
names.  PID has `SP : REAL` (setpoint); the moment generated_debug.cpp
references `g_config.INSTANCE0.PID1.SP`, the preprocessor turns `.SP`
into `.(*(volatile uint16_t *)(0x3D))` and avr-g++ fails with
`error: expected unqualified-id before '(' token`.  Any project with a
PID block on an AVR target (the OpenPLC Simulator board target as
well as real Arduino-family runtime-v3 boards) hits this on every
build.  Reported against rc1 strucpp 0.4.11.

Fix is a structural one — keep `<avr/io.h>` out of the TU that names
user variables:

  - new `runtime/include/debug_table.hpp` carries the AVR-clean
    subset generated_debug.cpp needs (Entry struct, TypeTag enum,
    STRUCPP_DEBUG_FLASH attribute).  Defines PROGMEM placement
    directly as `__attribute__((__progmem__))` so the pgmspace.h
    include can stay out.
  - debug_dispatch.hpp now `#include "debug_table.hpp"` for the
    shared ABI types (single source of truth, no drift between the
    table emit side and the runtime dispatch side) and keeps its own
    `<avr/pgmspace.h>` include for the inline `read_entry` / dispatch
    helpers that genuinely need `pgm_read_*_far`.  Only the runtime
    translation unit (arduino_runtime_glue.cpp / runtime_v4_entry.cpp)
    consumes that header; neither names user variables.
  - debug-table-gen.ts emits `#include "debug_table.hpp"` instead of
    `#include "debug_dispatch.hpp"` at the top of generated_debug.cpp.

End-to-end verification (atmega2560, same flags openplc-editor's
Simulator board uses):

  Pass A — new debug_table.hpp:    generated_debug.cpp compiles ✓
  Pass B — old debug_dispatch.hpp: fails with `error: expected
                                    unqualified-id before '(' token`
                                    on `.SP` (the original bug) ✓

Test suite: 1965 passed (was 1964 + the now-updated debug-table-gen
fixture asserting the new include).
